Update Partay.java
I need to set the array list to the tables and seats and the people with their order into tables and seats. Once this is done I will allow for people to be added into the array list to add people in the tables and seats if there are left over spots. I have already created the user prompts to ask to add people but I am struggling with the part of setting the array list and splitting all the information that is put into from the .text file. I will get a lot done over the weekend since I will be in West Virginia.